# docs/swagger.yaml
definitions:
...
  models.Product:
    properties:
      buy_permissions:
        items:
          type: number
        type: array
      contract_address:
        type: string
      contract_chain:
        type: string
      contract_type:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      preview_image:
        type: string
      price:
        type: number
      status:
        type: boolean
      stock:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  status.Response:
    properties:
      status_code:
        type: integer
      status_message:
        type: string
    type: object
  status.ResponseWtihData:
    properties:
      data:
        type: any
      status:
        $ref: '#/definitions/status.Response'
    type: object
info:
  contact:
    email: service@invar.finance
    name: InVar
    url: https://www.invar.finance/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /admin/product:
    get:
      consumes:
      - application/json
      description: 獲得全部商品清單
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Product'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得全部商品清單
      tags:
      - Product
    patch:
      consumes:
      - multipart/form-data
      description: 建立新商品
      parameters:
      - default: 1
        description: 庫存數量
        in: formData
        name: stock
        type: integer
      - description: 商品狀態
        in: formData
        name: status
        type: boolean
      - description: 商品價格
        in: formData
        name: price
        type: integer
      - description: 商品標題
        in: formData
        name: title
        type: string
      - description: 商品描述
        in: formData
        name: description
        type: string
      - description: 區塊鏈名稱
        in: formData
        name: contract_chain
        type: string
      - description: 合約類型
        in: formData
        name: contract_type
        type: string
      - description: 合約地址
        in: formData
        name: contract_address
        type: string
      - description: 購買權限
        in: formData
        items:
          type: integer
        name: buy_remissions
        type: array
      - description: 預覽檔案
        in: formData
        name: small_image
        type: file
      - description: 完整檔案
        in: formData
        name: large_image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "417":
          description: Expectation Failed
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 建立新商品
      tags:
      - Product
    post:
      consumes:
      - multipart/form-data
      description: 建立新商品
      parameters:
      - default: 1
        description: 庫存數量
        in: formData
        name: stock
        required: true
        type: integer
      - description: 商品狀態
        in: formData
        name: status
        required: true
        type: boolean
      - description: 商品價格
        in: formData
        name: price
        required: true
        type: integer
      - description: 商品標題
        in: formData
        name: title
        required: true
        type: string
      - description: 商品描述
        in: formData
        name: description
        required: true
        type: string
      - description: 區塊鏈名稱
        in: formData
        name: contract_chain
        required: true
        type: string
      - description: 合約類型
        in: formData
        name: contract_type
        required: true
        type: string
      - description: 合約地址
        in: formData
        name: contract_address
        required: true
        type: string
      - description: 購買權限
        in: formData
        items:
          type: integer
        name: buy_remissions
        required: true
        type: array
      - description: 預覽檔案
        in: formData
        name: small_image
        required: true
        type: file
      - description: 完整檔案
        in: formData
        name: large_image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "417":
          description: Expectation Failed
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 建立新商品
      tags:
      - Product
  /admin/product/{id}:
    get:
      consumes:
      - application/json
      description: 獲得全部商品清單
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得全部商品清單
      tags:
      - Product
...
  /product:
    get:
      consumes:
      - application/json
      description: 獲得全部商品清單
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Product'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得全部商品清單
      tags:
      - Product
  /product/{id}:
    get:
      consumes:
      - application/json
      description: 獲得全部商品清單
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得全部商品清單
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"